examples/TreeView: 将文件列表视图封装在SimpleFileView类中。
